backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
supabase>=2.5.0
pydantic>=2.7.0
pydantic-settings
boto3>=1.34.0
bcrypt
PyJWT
python-multipart
python-dotenv
httpx
requests
email-validator
